+++ lib/rails_webhook_outbox/testing.rb
@@ -1,0 +1,13 @@
+module RailsWebhookOutbox
+  module Testing
+    class << self
+      def deliveries
+        @deliveries ||= []
+      end
+
+      def clear_deliveries!
+        @deliveries = []
+      end
+    end
+  end
+end